requirements.txt
streamlit
google-generativeai
google.cloud
